+++ src/components/Auth/FamilyLoginScreen.jsx
@@ -25,6 +25,7 @@ import {
 } from '../../utils/imageFallback';
 import { GITHUB_REPOSITORY_URL } from '../../constants/project';
 import AndroidAppDownload from './AndroidAppDownload';
+import ReleasePreviewCard from './ReleasePreviewCard';
 import LanguageSwitcher from '../LanguageSwitcher';
 import ProjectSupportCard from '../ProjectSupportCard';
 import { PRODUCT_NAME } from '../../../shared/brand.js';
@@ -451,6 +452,7 @@ export default function FamilyLoginScreen({ onStartOnboarding, onOpenServerConfi
             )}
           </div>
         </div>
+        <ReleasePreviewCard />
         <AndroidAppDownload />
         </div>
       </main>